+++ dash/_no_update.py
@@ -4,6 +4,8 @@ class NoUpdate:
 
     @staticmethod
     def is_no_update(obj):
-        return isinstance(obj, NoUpdate) or (
-            isinstance(obj, dict) and obj == {"_dash_no_update": "_dash_no_update"}
+        return (
+            obj is NoUpdate
+            or isinstance(obj, NoUpdate)
+            or (isinstance(obj, dict) and obj == {"_dash_no_update": "_dash_no_update"})
         )